[SOURCEDEPS] Add script for generation of sourcedep debs
sourcedep debs are Debian binary packages usable for tracking
build depedencies. This is really a work-around for missing
functionality in APT. The different properties of each
source package are mapped to binary packages:

 package name -> name with appended '-build-depends'
 Build-Depends -> Depends
 Build-Depends-Indep -> Recommends
 Binary -> Suggest and Provides (in the Provides with appended
           '-build-depends')

The binary deps are arch independent if the Build-Depends contain no
arch restrictions, and arch dependent otherwise.

Added this in the packages.d.o code since it profits from the
archive retrival functionality.