Rework newpkg stuff
Create lists for all suites, not just sid (this also avoids
having to hardcode the suite name, which is nicer for the
Ubuntu archive).
Handle packages correctly that get added and immediatly deleted
or deleted and immediatly readded. Not that I would have some
examples for that but it might happen.