Disable gettext translations
mpm_worker, mod_perl and gettext together show some funny
results. When several clients request different languages from
the server simultaniously they can get the completly wrong or even
mixed strings. Which of course doesn't happen on my test server,
since the chances for that to happen are very slim there.

This whole mess can easily explained by the fact that the locale
setting done with setlocale is global to the process. The solution
for that is most likely to switch to a OO aproach with using a
message catalog (no idea how to tell sprintf which locale to use
in a non-global way, though...)