Various improvements to the search result comments
Don't pretend we know how many results have been suppressed,
we really have no idea (that's the whole point of short cutting
the search).

Don't declare we will display results when we wont. We have all
the necessary information so use it.

Add some useful links similar to what is on the search_contents
site.