Separate handling of po translations and DDTP translations
This way we don't need to immediatly fall back to English
if people have more than one preferred language for which
only part of the page content is available.

If e.g. they would have "ca;q=0.9, de;q=0.5, en;q=0.1" in
their Accept-Language header (insert any DDTP-only language
for ca and any PO-available language for de), we can show the
description in Catalan if available, and the po contents in
German.

The code will usually try to fall back to $po_lang if a
description is not available in $ddtp_lang.

Not implemented (yet) is a DDTP language negotiation _after_
we actually know which DDTP translations are available.

The 'lang' query parameter overrides both values and has
only the fallback English (But you shouldn't use it
anyway ;).