[UBUNTU] Packages::DoSearchContents: Don't die when no information is available
Issue a real error message if no contents database is available for the suite
in question.

This isn't exposed on the Debian site since we actually have _some_ contents
for every suite (experimental has contents for debian-ports.org). But on
the Ubuntu site we have no information at all for *-backports and *-updates.