debtags-xgettext: Handle linebreaks (and other whitespace) a little more sanely
Do not include linebreaks from the vocabulary file in the msgids.
Replace \n.\n with a single linebreak. Remove the single whitespace
at the beginning of the long description and new paragraphs.

Together these changes should make it easier for translators to translate
the long descriptions.

fix_old_debtags_po: Script to convert the old .po files to the new format
so that translators don't have to do that manually.