create_index_pages: Do not die due to namespace collisions
If for whatever reason the same name appears in more than one
set of sections, subsections, and priorities, the code used
to die. The new code will not handle this very well either
(you will get repeated lines in the affected file), but
at least it will not take everything else with it.