Packages::DoShow: Don't fall back to "us" archive to find source
If the archive really doesn't have source (debports
is the usual offender here because it doesn't have any
sources and sometimes doesn't delete stale packages)
don't try to point to some source in the main archive
(e.g. of a newer source version which doesn't build this
binary package anymore). This is just confusing. Also this
generated source links to the debports FTP which is wrong
(again, there is no source there).
Closes: #684118