Allow Contents files be symlinks
The safety guard in this was last used in wheezy which is discontinued.
On the other hand this hindered to update the filelist for newer suites
because they got switched to symlinks to by-hash files and thus all of
them now are symlinks.  We removed wheezy (not supported anymore) so we
can also remove that symlink check for duplications.

Closes: way too many duplicated bugs that piled up over time :/