Concatenate results from Contents-${arch}.gz and Contents-all.gz when the latter is present. (Closes: #977006, #977743, #980888)
Since commit 81824d23 in daklib, the archive provides separate Contents files
with packages from Arch:all for some suites. [1] (All suites post-buster appear to
be affected.) Those file lists were not scanned for the web pages powering
packages.d.o, and the file lists were unavailable online.

This commit attempts to resolve the issue by parsing the file for Arch:all
immediately afterwards, if it is present. The results are effectively
concatenated.

A better long-term solution might be to produce separate transfer files for
Arch:all, but that may not work until buster is dropped from the archive.

[1] https://salsa.debian.org/ftp-team/dak/-/commit/81824d2326f5cc50fdcb95c81f9f26864aebaa15